+++ src/grid/grid-filtering-ui.spec.ts
@@ -502,7 +502,7 @@ describe("IgxGrid - Filtering actions", () => {
             verifyFilterUIPosition(filterUIContainer, grid);
 
             // false condition
-            options[1].selected = true;
+            options[2].selected = true;
             select.nativeElement.dispatchEvent(new Event("change"));
             fix.detectChanges();
             expect(grid.rowList.length).toEqual(2);
@@ -512,7 +512,7 @@ describe("IgxGrid - Filtering actions", () => {
             expect(reset.nativeElement.classList.contains("igx-button--disabled")).toBeFalsy();
 
             // true condition
-            options[0].selected = true;
+            options[1].selected = true;
             select.nativeElement.dispatchEvent(new Event("change"));
             fix.detectChanges();
             expect(grid.rowList.length).toEqual(3);
@@ -522,8 +522,16 @@ describe("IgxGrid - Filtering actions", () => {
             expect(close.nativeElement.classList.contains("igx-button--disabled")).toBeFalsy();
             expect(reset.nativeElement.classList.contains("igx-button--disabled")).toBeFalsy();
 
+            // (all) condition
+            options[0].selected = true;
+            select.nativeElement.dispatchEvent(new Event("change"));
+            fix.detectChanges();
+            expect(grid.rowList.length).toEqual(8);
+            expect(close.nativeElement.classList.contains("igx-button--disabled")).toBeFalsy();
+            expect(reset.nativeElement.classList.contains("igx-button--disabled")).toBeTruthy();
+
             // null condition
-            options[2].selected = true;
+            options[3].selected = true;
             select.nativeElement.dispatchEvent(new Event("change"));
             fix.detectChanges();
             expect(grid.rowList.length).toEqual(2);
@@ -533,7 +541,7 @@ describe("IgxGrid - Filtering actions", () => {
             expect(reset.nativeElement.classList.contains("igx-button--disabled")).toBeFalsy();
 
             // not null condition
-            options[3].selected = true;
+            options[4].selected = true;
             select.nativeElement.dispatchEvent(new Event("change"));
             fix.detectChanges();
             expect(grid.rowList.length).toEqual(6);
@@ -547,7 +555,7 @@ describe("IgxGrid - Filtering actions", () => {
             expect(reset.nativeElement.classList.contains("igx-button--disabled")).toBeFalsy();
 
             // empty condition
-            options[4].selected = true;
+            options[5].selected = true;
             select.nativeElement.dispatchEvent(new Event("change"));
             fix.detectChanges();
             expect(grid.rowList.length).toEqual(3);
@@ -558,7 +566,7 @@ describe("IgxGrid - Filtering actions", () => {
             expect(reset.nativeElement.classList.contains("igx-button--disabled")).toBeFalsy();
 
             // not empty condition
-            options[5].selected = true;
+            options[6].selected = true;
             select.nativeElement.dispatchEvent(new Event("change"));
             fix.detectChanges();
             expect(grid.rowList.length).toEqual(5);