+++ src/grid/grid-filtering.component.ts
@@ -106,6 +106,7 @@ export class IgxGridFilterComponent implements IGridBus, OnInit, OnDestroy, DoCh
         return `igx-filtering`;
     }
 
+    public booleanFilterAll = "All";
     public dialogShowing = false;
     public dialogPosition = "igx-filtering__options--to-right";
 
@@ -133,6 +134,9 @@ export class IgxGridFilterComponent implements IGridBus, OnInit, OnDestroy, DoCh
 
     @ViewChild("select", { read: ElementRef})
     protected select: ElementRef;
+
+    @ViewChild("input", { read: ElementRef})
+    protected input: ElementRef;
 
     constructor(private zone: NgZone, public gridAPI: IgxGridAPIService, public cdr: ChangeDetectorRef, private elementRef: ElementRef) {
         this.filterChanged.pipe(
@@ -167,6 +171,9 @@ export class IgxGridFilterComponent implements IGridBus, OnInit, OnDestroy, DoCh
     public refresh() {
         this.dialogShowing = !this.dialogShowing;
         if (this.dialogShowing) {
+            if(this.input) {
+                this.input.nativeElement.focus();
+            }
             this.column.filteringCondition = this.getCondition(this.select.nativeElement.value);
         }
         this.cdr.detectChanges();
@@ -213,6 +220,10 @@ export class IgxGridFilterComponent implements IGridBus, OnInit, OnDestroy, DoCh
     }
 
     public selectionChanged(value): void {
+        if (value === this.booleanFilterAll) {
+            this.clearFiltering(true);
+            return;
+        }
         this._filterCondition = value;
         this.column.filteringCondition = this.getCondition(value);
         if (this.unaryCondition) {