+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 All notable changes for each version of this project will be documented in this file.
 ## 6.0.0
+- Theming - You can now use css variables to style the component instances you include in your project.
 - Added `onDoubleClick` output to `igxGrid` to emit the double clicked cell.
 - Added `findNext`, `findPrev` and `clearSearch` methods to the IgxGridComponent which allow easy search of the grid data, even when the grid is virtualized.
 - Added `IgxTextHighlightDirective` which highlights parts of a DOM element and keeps and updates "active" highlight.