+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes for each version of this project will be documented in this 
 - Added `onDoubleClick` output to `igxGrid` to emit the double clicked cell.
 - Added `findNext`, `findPrev` and `clearSearch` methods to the IgxGridComponent which allow easy search of the grid data, even when the grid is virtualized.
 - Added `IgxTextHighlightDirective` which highlights parts of a DOM element and keeps and updates "active" highlight.
+- Added `All` option to the filter UI select for boolean columns
+
 ## 5.3.1
 - igx-dialog changes
     - Dialog title as well as dialog actions (buttons) can be customized. For more information navigate to the [ReadMe](https://github.com/IgniteUI/igniteui-angular/blob/master/src/dialog/README.md).