+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes for each version of this project will be documented in this 
 ## 6.0.0
 - Theming - You can now use css variables to style the component instances you include in your project.
 - Added `onDoubleClick` output to `igxGrid` to emit the double clicked cell.
+- Added `All` option to the filter UI select for boolean columns
 
 ## 5.3.1
 - igx-dialog changes