+++ CHANGELOG.md
@@ -1,12 +1,16 @@
 # Ignite UI for Angular Change Log
 
 All notable changes for each version of this project will be documented in this file.
+## 5.3.2
+-  Added "All" option in the filtering UI for a boolean column 
+
 ## 5.3.1
 - igx-dialog changes
     - Dialog title as well as dialog actions (buttons) can be customized. For more information navigate to the [ReadMe](https://github.com/IgniteUI/igniteui-angular/blob/master/src/dialog/README.md).
-- Filtering a boolean column by `false` condition will return only the real `false` values, excluding `null` and `undefined`. Filtering by `Null` will return the `null` values and filtering by `Empty` will return the `undefined`.
-- The `Filter` button in the filtering UI is replaced with a `Close` button that is always active and closes the UI.
-- Filtering UI input displays a `X` icon that clears the input.
+- igx-grid changes 
+    - Filtering a boolean column by `false` condition will return only the real `false` values, excluding `null` and `undefined`. Filtering by `Null` will return the `null` values and filtering by `Empty` will return the `undefined`.
+    - The `Filter` button in the filtering UI is replaced with a `Close` button that is always active and closes the UI.
+    - Filtering UI input displays a `X` icon that clears the input.
 - [#1136](https://github.com/IgniteUI/igniteui-angular/issues/1136) Exception is thrown when trying to pin column in certain scenarios
 - [#1124](https://github.com/IgniteUI/igniteui-angular/issues/1124) Resizing a column to where horizontal scrollbar should be shown doesn't show the scrollbar
 - [#1169](https://github.com/IgniteUI/igniteui-angular/issues/1169) Misalignment when resize a column and then scroll horizontally